feat: Alias copying pwd to Windows temp from WSL
Adds alias for copying the current directory to a new temp dir
from WSL. This is a little hack to work around limitations with
running code in powershell from the WSL directory. It's often
preferrable to work from the WSL filesystem (Windows kept
corrupting my git indexes) but this makes execution from Windows
tricky in the infrequent cases where something needs to be tested
from PowerShell.